remove student home  component
